initial commit of file
Key fields are:  CITY_MARKET_ID, which is the same for all the NYC airports, for example.   and CITY_MARKET_NAME_FULL.    I think the CITY_MARKET_NAME_FULL list could be the right list for target destinations.  then we grab all CITY_MARKET_ID for that metro area and all the airports with that CITY_MARKET_ID .   many of these airport will have no flights of interest to us.  For example, in Seattle, Boeing has it’s own airport, which is of no interest to our project…